Internal DBMS Version number adjustment from MMMM SSS PP (1570 003 05) to MM MM MM SSSS PPPP (15 07 00 0003 0005) This accommodates for 2 number Minor/Maintenance releases (Postgres and MySQL) And larger SupportPackages / PatchLevels (which can now hold a typical build number of 4 numbers)

+++ asetune/trunk/src/com/asetune/ICounterController.java
@@ -145,7 +145,7 @@ public interface ICounterController
 	Timestamp getStatisticsLastSampleTime();
 	void resetStatisticsTime();
 
-	void initCounters(DbxConnection conn, boolean b, int srvExecutableVersionNum, boolean clusterEnabled, int mdaVersion) throws Exception;
+	void initCounters(DbxConnection conn, boolean b, long srvExecutableVersionNum, boolean clusterEnabled, long mdaVersion) throws Exception;
 
 	void reset(boolean b);
 
@@ -154,7 +154,7 @@ public interface ICounterController
 
 
 	
-	List<CountersModel> getCmListDependsOnConfig(String srvConfig, DbxConnection conn, int srvVersionNum, boolean isClusterEnabled);
+	List<CountersModel> getCmListDependsOnConfig(String srvConfig, DbxConnection conn, long srvVersionNum, boolean isClusterEnabled);
 	void setWaitEvent(String string);
 	void createCounters(boolean hasGui);
 	boolean sleep(int i);
