+++ asetune/trunk/src/com/asetune/cm/rs/CmAdminStats.java
@@ -45,8 +45,8 @@ extends CountersModel
 	public static final String   GROUP_NAME       = MainFrameRs.TCP_GROUP_MC;
 	public static final String   GUI_ICON_FILE    = "images/"+CM_NAME+".png";
 
-	public static final int      NEED_SRV_VERSION = 0;
-	public static final int      NEED_CE_VERSION  = 0;
+	public static final long     NEED_SRV_VERSION = 0;
+	public static final long     NEED_CE_VERSION  = 0;
 
 	public static final String[] MON_TABLES       = new String[] {"stats"};
 	public static final String[] NEED_ROLES       = new String[] {};
@@ -163,13 +163,13 @@ extends CountersModel
 	}
 
 	@Override
-	public String[] getDependsOnConfigForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String[] getDependsOnConfigForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		return NEED_CONFIG;
 	}
 
 	@Override
-	public void addMonTableDictForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public void addMonTableDictForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		try 
 		{
@@ -192,7 +192,7 @@ extends CountersModel
 	}
 
 	@Override
-	public List<String> getPkForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public List<String> getPkForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		List <String> pkCols = new LinkedList<String>();
 
@@ -245,7 +245,7 @@ extends CountersModel
 	}
 
 	@Override
-	public String getSqlForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String getSqlForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		boolean sample_resetAfter = Configuration.getCombinedConfiguration().getBooleanProperty(PROPKEY_sample_resetAfter, DEFAULT_sample_resetAfter);
 
@@ -260,7 +260,7 @@ extends CountersModel
 	}
 	
 	@Override
-	public String getSqlInitForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String getSqlInitForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		String statsOn = 
 			  "--admin stats, cancel -- cancel might crach RS in some cases...\n"
@@ -282,7 +282,7 @@ String AOBJ = "trace 'on', 'dsi', 'dsi_workload' \n" + "go \n";
 	}
 
 //	@Override
-//	public String getSqlCloseForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+//	public String getSqlCloseForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 //	{
 //		String statsOff = 
 //			  "--admin stats, cancel -- cancel might crach RS in some cases...\n"
