+++ asetune/trunk/src/com/asetune/cm/rs/CmRsSrcToDest.java
@@ -69,8 +69,8 @@ extends CountersModel
 	public static final String   GROUP_NAME       = MainFrameRs.TCP_GROUP_MC;
 	public static final String   GUI_ICON_FILE    = "images/"+CM_NAME+".png";
 
-	public static final int      NEED_SRV_VERSION = 0;
-	public static final int      NEED_CE_VERSION  = 0;
+	public static final long     NEED_SRV_VERSION = 0;
+	public static final long     NEED_CE_VERSION  = 0;
 
 	public static final String[] MON_TABLES       = new String[] {"stats"};
 	public static final String[] NEED_ROLES       = new String[] {};
@@ -272,19 +272,19 @@ extends CountersModel
 	}
 
 	@Override
-	public String[] getDependsOnConfigForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String[] getDependsOnConfigForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		return NEED_CONFIG;
 	}
 
 	@Override
-	public String getSqlForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String getSqlForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		return "";
 	}
 
 	@Override
-	public List<String> getPkForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public List<String> getPkForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		List <String> pkCols = new LinkedList<String>();
 
@@ -295,7 +295,7 @@ extends CountersModel
 	}
 
 //	@Override
-//	public String getSqlForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+//	public String getSqlForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 //	{
 //		return "";
 //	}
