+++ asetune/trunk/src/com/asetune/sql/conn/IqConnection.java
@@ -46,7 +46,7 @@ extends TdsConnection
 	}
 
 	@Override
-	public int getDbmsVersionNumber()
+	public long getDbmsVersionNumber()
 	{
 //		return AseConnectionUtils.getAseVersionNumber(this);
 		return AseConnectionUtils.getIqVersionNumber(this);
